tests: log_test_mock.sh: allow for selection of particular test
This is to cut down some time when only particular item of the
test matrix is of interest, such as in

  ./log_test_mock.sh -v -ncl -ncsc -tqb+_il+_c+ <srpm>

command that can be used to highlight the issue with binutils 2.29+
compatibility fix making for a new issue when all of following are
satisfied:

- there are three link participants (reusing terminology establish
  in the patches constituting said fix)
  client program > intermediate library > libqb
  (client is not necessarily directly dependent on libqb)

- client program is build-time linked using ld/binutils either
  < or >= 2.29 but with libqb 1.0.3 (containing the said fix,
  i.e., new linker script is in effect for building the client
  program)

- client program does not use any logging at all

- intermediate library triggers some checks when gets loaded,
  through usage of QB_LOG_INIT_DATA macro from qb/qblog.h,
  which are meant to check itself, but alas, the symbols to
  check on behalf of this library get run-time overridden
  with symbols of the client program, which only do exist
  because the libqb.so linker script (intended for other
  effect) made it happen despite the section is empty

So far, log_test_mock.sh was only thoroughly run without any
log-suppressing options (e.g. -ncl), which together with some
extra linker/build flags (distro specific?) in case of compiling
pacemaker (exercising the same scheme with some CLI utilities)
masked this problem.  [this all is still to be verified]

This asks for a follow-up solution, and this commit is a good
start as running whole matrix would keep increasing turnaround
time needlessly (the new skipping harness is not ideal, but
serves the purpose).

Signed-off-by: Jan Pokorný <jpokorny@redhat.com>