Low: qb_{sigpipe_ctl,socket_nosigpipe}: optimize through inlining
It was pointed out by Chrissie that the former function, qb_sigpipe_ctl,
can be a bottle neck when SIGPIPE handling is being temporarily
ignored around UNIX socket IO (which also makes libqb generally
thread-unfriendly!), unless the platform provides better means of
achieving that effect, in which case the function is no-op.
Due to multiple callers, it would be impractical to have such
a build-time conditionalization at the callsites rather than at that
function itself, but with Linux and FreeBSD (amongst possibly more)
this leads to entering completely empty frames needlessly.  Solution is
to treat the function as inline at the affected callsites, which feeds
the compiler's optimizing phases as expected.  Nonetheless, to retain
(API-bypassing) ABI compatibility, we also preserve said function
+ another falling into the same category, qb_socket_nosigpipe,
also as externally visible symbols.

Signed-off-by: Jan Pokorný <jpokorny@redhat.com>