log: journal: implicitly carry just pristine (redundancy-free) message
All relevant metadata incl. priority that gets normally exposed
otherwise is stored in the per-message payload in the journal,
where it can be (message-filtering-)fetched from if needed
(which shall not be needed most of time, anyway, since the priority
is already reflected in the line formatting (red/black/bold)
in the interactive journal browsing session.

Of course, users are welcome to set their own format string
(before or after turning syslog target into journal one),
even setting NULL format string afterwards will reset it
to the global implicit default.

[
I'd tend to do the same for casual syslog and blackbox logging,
since it is pretty much in the same boat (priority shall be _metadata_
rather then plaintext in-message indicator), but it would touch
the assumed behaviour consistency promises -- of which there are
luckily none for the journal output, yet.
]

Also finalize, extend and proofread the comment for qb_log_format_set.

Signed-off-by: Jan Pokorný <jpokorny@redhat.com>