ipc: Use mkdtemp for more secure IPC files
This is not meant to be the last word on the subject, it's mainly a
simple way of making the current libqb more secure. It's backwards
compatible with an old server.

It calls rmdir on the directory created by mkdtemp *way* too often, but
it seems to be the only way to be sure that things get cleaned up on
the various types of server/client exit. I'm sure we can come up with
something tidier for master but I hope this, or something similar, will
be OK for 1.0.x.

The O_EXCL patch is also still recommended, of course.