log: fix potential overflow with long log messages
qb_vsnprintf_serialize was called with 'max_size' as the
limiting number for the length of the formatted log
message. But the buffer also needs to contain the
log header (given by 'actual_size'), so we now pass
'max_size - actual_size' as the maximum length of the
formatted log message.

Also added error checks to the blacbkbox calls at
the end of the test, as these now provide a proper
test that the BB is functioning. Before they were
masking failures.